requirements.txt
streamlit
python-pptx
google-generativeai
transformers
huggingface-hub
python-dotenv
tqdm
